requirements.txt
numpy==2.1.2
pandas==2.2.3
Flask
openai 
pdfplumber
python-dotenv